These brushes were both coded by me, the Mixer brush being my first one and the Tree Remover my second.
What the Tree Remover does is simply remove trees. You select the dimensions of the area you want to remove trees from by using the command: "/b tr dim [x] [y] [z]". So for example "/b tr dim 5 2 5" will set the height to 2 and the length and width to 5. When you set the dimensions you simply right click with the feather or gunpowder and it will remove all trees within that area. Also when a tree is removed, the dirt under the bottom log will be changed to grass so that you dont have to go and replace all of the dirt with grass.

What the Mixer brush does is it mixes a block of your choice by using the command "/v [item]" in with a certain area of where you clicked. So in this brush you also set the dimensions by doing "/b mixer dim [x] [y] [z]"
and it will set the area you want to edit when you click. So for example "/b tr dim 5 2 5" will set the height to 2 and the length and width to 5. The when you right click with the feather or gunpowder, inside of the area, it randomly selects blocks to set the block of your choice to. So it really just mixes land of your choice with a block of your choice and it's all random so it will never look the same.